Recall Bookmark
A Boolmark recall now sets the frequency offset to zero and thus the bookmark frequency = hardware frequency. If e.g. the offset frequency is 700 Khz and the bookmark has a frequency below that, it could not be set from a bookmark up to now.

Then void MainWindow :: setPassband has been changed so that asymmetrical filter bandwidths also work as with the change in void MainWindow::onBookmarkActivated