Another small change in the behavior of the user bandwidth...
User bandwidth is retained when the demodulator is not changed. This is both if you accidentally click mode and stick to the same demodulator, and when selecting a bookmark if it uses the same demodulator.

Probably not enough to the requirements of a professional software, but if someone is interested, he can install it by himself...